Extract query class for Lock.
