+++ opengever/locking/model/__init__.py
@@ -1,5 +1,5 @@
-from opengever.locking.model.locks import Lock
-
+from opengever.locking.model.locks import Lock  # noqa
+from opengever.locking.model import query  # noqa, keep, initializes query classes!
 
 tables = [
     'locks',
