+++ opengever/locking/model/locks.py
@@ -3,7 +3,6 @@ from opengever.base.date_time import utcnow_tz_aware
 from opengever.base.model import Base
 from opengever.base.model import UTCDateTime
 from opengever.ogds.models import USER_ID_LENGTH
-from opengever.ogds.models.query import BaseQuery
 from sqlalchemy import Column
 from sqlalchemy import Integer
 from sqlalchemy import String
@@ -18,23 +17,10 @@ def lowest_valid():
     return utcnow_tz_aware() - timedelta(seconds=DEFAULTTIMEOUT)
 
 
-class LockQuery(BaseQuery):
-
-    def valid_locks(self, object_type, object_id):
-        query = Lock.query.filter_by(object_type=object_type,
-                                     object_id=object_id)
-        return query.filter(Lock.time >= lowest_valid())
-
-    def invalid_locks(self):
-        return Lock.query.filter(Lock.time < lowest_valid())
-
-
 class Lock(Base):
 
     __tablename__ = 'locks'
     __table_args__ = (UniqueConstraint('object_id', 'object_type', 'lock_type'), {})
-
-    query_cls = LockQuery
 
     lock_id = Column("id", Integer, Sequence("locks_id_seq"), primary_key=True)
     object_id = Column(Integer)
